requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-cors-headers==3.7.0
django-environ==0.4.5
django-extensions==3.1.1
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.12.0
djangorestframework==3.12.2
drf-yasg==1.20.0
et-xmlfile==1.0.1
flake8==3.9.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mccabe==0.6.1
openpyxl==3.0.7
packaging==20.9
pycodestyle==2.7.0
pyflakes==2.3.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
#Para CircleCI
psycopg2==2.8.6